app/requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.9
Flask-HTTPAuth==4.5.0
Werkzeug==2.0.2
gunicorn==20.1.0
Click==7.1.2
flake8==4.0.1
python-dotenv==0.19.2
python-json-logger==2.0.2
requests==2.27.1
setuptools==59.6.0
redis==4.1.2
celery==5.1.2
Flask-SQLAlchemy==2.5.1